requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.29.0,<1.0.0
pydantic>=2.7.0,<3.0.0
pyyaml>=6.0.1,<7.0.0
mcp[cli]>=1.0.0,<2.0.0
selenium>=4.20.0,<5.0.0
websockets>=12.0,<14.0.0
aiofiles>=23.2.0,<25.0.0
httpx>=0.27.0,<1.0.0
